checker ok

+++ utils/ft_atoi.c
@@ -3,16 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   ft_atoi.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mac <mac@student.42.fr>                    +#+  +:+       +#+        */
+/*   By: fnaciri- <fnaciri-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/04/18 11:13:37 by mac               #+#    #+#             */
-/*   Updated: 2021/04/18 11:13:52 by mac              ###   ########.fr       */
+/*   Updated: 2021/04/18 14:49:44 by fnaciri-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../include/utils.h"
+#include "../include/checker.h"
 
-int		ft_atoi(const char *s)
+long		ft_atoi(const char *s)
 {
 	int		i;
 	size_t	n;
